# ApiGateway/app/openapi.yaml
openapi: 3.0.0
info:
  title: Hệ thống Quản lý Thanh toán Học phí (Microservices SOA)
  version: 1.0.0
  description: |
    Tài liệu API cho hệ thống Microservices bao gồm Auth, Customer Management, Payment Processor và Tuition Service.
    
    Tất cả các request **ngoại trừ /auth/login và /auth/logout** đều yêu cầu xác thực bằng **JWT Cookie** có tên là `jwt_token`. Gateway sẽ tự động trích xuất `customerId` từ JWT và thêm vào header `X-Customer-Id` trước khi gửi đến các dịch vụ backend.
servers:
  - url: http://localhost:8080 # Địa chỉ mặc định của API Gateway
security:
  - jwtCookie: [] # Áp dụng bảo mật cho các endpoint không phải là login/logout
...
paths:
...
  /tuition/studentId/{studentId}:
    get:
      tags:
        - Tuition Service
      summary: Lấy danh sách các khoản học phí theo ID sinh viên.
      operationId: getTuitionsByStudentId
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
          description: ID sinh viên (ví dụ 523H1101).
      responses:
        '200':
          description: Lấy danh sách học phí thành công.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/StudentTuition'
              example:
                success: true
                message: Student tuitions retrieved successfully.
                data:
                  - tuitionId: 1
                    studentId: 523H1101
                    studentName: Nguyễn Văn A
                    amount: 2800000
                    semester: "I"
                    academicYear: 2023-2024
                    isPaid: false
                  - tuitionId: 2
                    studentId: 523H1101
                    studentName: Nguyễn Văn A
                    amount: 3500000
                    semester: "II"
                    academicYear: 2023-2024
                    isPaid: true
        '400':
          description: ID sinh viên không hợp lệ.
          $ref: '#/components/responses/BadRequestGeneric'
        '404':
          description: Không tìm thấy hồ sơ học phí nào cho ID sinh viên này.
          $ref: '#/components/responses/NotFoundGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tuition/id/{tuitionId}:
    get:
      tags:
        - Tuition Service
      summary: Lấy chi tiết một khoản học phí theo ID học phí.
      operationId: getTuitionById
      parameters:
        - name: tuitionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID của khoản học phí.
      responses:
        '200':
          description: Lấy chi tiết học phí thành công.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/StudentTuition'
              example:
                success: true
                message: Tuition retrieved successfully.
                data:
                  tuitionId: 1
                  studentId: 523H1101
                  studentName: Nguyễn Văn A
                  amount: 2800000
                  semester: "I"
                  academicYear: 2023-2024
                  isPaid: false
        '400':
          description: ID học phí không hợp lệ.
          $ref: '#/components/responses/BadRequestGeneric'
        '404':
          description: Không tìm thấy hồ sơ học phí.
          $ref: '#/components/responses/NotFoundGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tuition/status:
    post:
      tags:
        - Tuition Service
      summary: Cập nhật trạng thái thanh toán của một khoản học phí. (Chỉ dùng nội bộ bởi Payment Processor)
      operationId: updateTuitionStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TuitionPaymentRequest'
      responses:
        '200':
          description: Cập nhật trạng thái thành công.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GenericResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/StudentTuition'
              example:
                success: true
                message: Tuition status updated to PAID successfully.
                data:
                  tuitionId: 1
                  studentId: 523H1101
                  studentName: Nguyễn Văn A
                  amount: 2800000
                  semester: "I"
                  academicYear: 2023-2024
                  isPaid: true
        '400':
          description: Lỗi dữ liệu request (thiếu trường, sai định dạng).
          $ref: '#/components/responses/BadRequestGeneric'
        '403':
          description: Lỗi nghiệp vụ (ví dụ Học phí không thuộc về sinh viên này).
          $ref: '#/components/responses/ForbiddenGeneric'
        '404':
          description: Không tìm thấy hồ sơ học phí.
          $ref: '#/components/responses/NotFoundGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
          
  # --------------------------------------------------------
  # PAYMENT PROCESSOR SERVICE ENDPOINTS
  # --------------------------------------------------------
...
components:
  schemas:
...
    GenericResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Cho biết request có thành công hay không.
          example: true
        message:
          type: string
          description: Thông báo mô tả kết quả.
          example: Current balance retrieved successfully.
        data:
          description: Dữ liệu trả về (kiểu dữ liệu cụ thể phụ thuộc vào endpoint).
          nullable: true
          example: null
...
    StudentTuition:
      type: object
      properties:
        tuitionId:
          type: integer
          format: int64
          example: 1
        studentId:
          type: string
          example: 523H1101
        studentName:
          type: string
          example: Nguyễn Văn A
        amount:
          type: integer
          format: int32
          example: 2800000
        semester:
          type: string
          example: "I"
        academicYear:
          type: string
          example: 2023-2024
        isPaid:
          type: boolean
          example: false
          description: Trạng thái thanh toán.
    TuitionPaymentRequest:
      type: object
      description: Request body để cập nhật trạng thái học phí.
      properties:
        tuitionId:
          type: integer
          format: int64
          example: 1
        isPaid:
          type: boolean
          example: true
      required:
        - tuitionId
        - isPaid
        
    # --------------------------------------------------------
    # PAYMENT PROCESSOR SCHEMAS
    # --------------------------------------------------------
...
  securitySchemes:
    jwtCookie:
      type: apiKey
      in: cookie
      name: jwt_token
      description: JWT token được trả về sau khi đăng nhập thành công. (HttpOnly Cookie)
  responses:
    Unauthorized:
      description: Yêu cầu xác thực. JWT Cookie không được tìm thấy hoặc không hợp lệ/hết hạn.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Authentication required. JWT Cookie not found.
    BadRequestGeneric:
      description: Lỗi nghiệp vụ hoặc lỗi dữ liệu (400 Bad Request).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
          example:
            success: false
            message: Insufficient balance to perform the debit. Required 2800000, available 1000000.
            data: null
    NotFoundGeneric:
      description: Không tìm thấy tài nguyên (404 Not Found).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
          example:
            success: false
            message: Customer with ID 999 not found.
            data: null
    ForbiddenGeneric:
      description: Không có quyền truy cập (403 Forbidden).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
          example:
            success: false
            message: You do not have permission to modify this tuition record.
            data: null
...